Only request a single RepoProvider in Notify task
Previously, the Notify task would get a repo provider token multiple times.
In the task itself, in the `ComparisonProxy`, and in each one of the configured checks.

This will now make sure to only get the repo provider once, and thread it through to all of its use-sites.
