+++ services/notification/notifiers/status/base.py
@@ -7,7 +7,6 @@ from shared.helpers.cache import NO_VALUE, make_hash_sha256
 from shared.torngit.base import TorngitBaseAdapter
 from shared.torngit.exceptions import TorngitClientError, TorngitError
 
-from database.models.core import Commit
 from helpers.cache import cache
 from helpers.match import match
 from services.comparison import ComparisonProxy, FilteredComparison
@@ -15,7 +14,6 @@ from services.notification.notifiers.base import (
     AbstractBaseNotifier,
     NotificationResult,
 )
-from services.repository import get_repo_provider_service_for_specific_commit
 from services.urls import get_commit_url, get_pull_url
 from services.yaml import read_yaml_field
 from services.yaml.reader import get_paths_from_flags
@@ -103,13 +101,6 @@ class StatusNotifier(AbstractBaseNotifier):
         return (
             len(report_uploaded_flags.intersection(flags_included_in_status_check)) > 0
         )
-
-    def repository_service(self, commit: Commit) -> TorngitBaseAdapter:
-        if not self._repository_service:
-            self._repository_service = get_repo_provider_service_for_specific_commit(
-                commit, fallback_installation_name=self.gh_installation_name
-            )
-        return self._repository_service
 
     def get_notifier_filters(self) -> dict:
         flag_list = self.notifier_yaml_settings.get("flags") or []
@@ -317,9 +308,9 @@ class StatusNotifier(AbstractBaseNotifier):
             )
 
     def send_notification(self, comparison: ComparisonProxy, payload):
+        repository_service = self.repository_service
         title = self.get_status_external_name()
         head_commit_sha = comparison.head.commit.commitid
-        repository_service = self.repository_service(comparison.head.commit)
         head_report = comparison.head.report
         state = payload["state"]
         message = payload["message"]
